添加接口功能，同时添加评论功能

+++ src/main/java/com/hl/community/model/Comment.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class Comment {
-    private Integer id;
+    private Long id;
     private Long parentId;
     private Integer type;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Integer likeCount;
